Update admin.py
Hi!
I've seen many django libraries (the last one with which I had issues was django-mptt) that handles django-admin's actions, expecting them to be an iterable object: (line 98 https://github.com/django-mptt/django-mptt/blob/master/mptt/admin.) I know that it's THEIR issue as Django Docs specifies that if someone wants to disable all actions, the best way is to set the actions attribute to None, and I am by the way addressing another issue on their project suggesting them to add the check "if actions is not None and ..." but I guess it could improve reliability and interoperability with other libraries to return an empty iterable. 

Greetings!